Chore: jump action implement!

+++ src/main/java/com/github/yoruhinda/aurakingdom/core/entity/Entity.java
@@ -3,8 +3,8 @@ package main.java.com.github.yoruhinda.aurakingdom.core.entity;
 import java.awt.*;
 
 public abstract class Entity {
-    protected int x;
-    protected int y;
+    protected float x;
+    protected float y;
 
     protected Entity(int x, int y){
         this.x = x;
@@ -19,19 +19,19 @@ public abstract class Entity {
 
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
